colaid_backend/requirements.txt
flask
opencv-python
numpy<2.0.0
pillow
flask-sqlalchemy
flask-login
python-dotenv
